Split ReferralService into draft and non-draft service
So that each distinct area has its own namespace, reducing clutter and
increasing cohesion

Also removes deprecated `legacyUpdateReferralDetails`
